FISH-10801 added the collection for the cpu_monitor.csv file for local log collection, the remote command collects the log automatically.

+++ appserver/extras/diagnostics-tool/src/main/java/fish/payara/extras/diagnostics/collection/collectors/LocalLogCollector.java
@@ -103,6 +103,7 @@ public class LocalLogCollector extends FileCollector {
             } else {
                 collectLogs(logPath, outputPath.resolve("logs"), logName);
             }
+            collectLogs(logPath, outputPath.resolve("logs"), "cpu_monitor");
         }
 
         return 0;
@@ -177,7 +178,11 @@ public class LocalLogCollector extends FileCollector {
             if (obfuscateEnabled && fileContains.equals("server.log")){
                 Obfuscation.obfuscateLogData(file, resolvedDestination);
             }else {
-                Files.copy(file, resolvedDestination);
+                if (Files.exists(resolvedDestination)) {
+                    LOGGER.info("File already exists, skipping: " + resolvedDestination);
+                } else {
+                    Files.copy(file, resolvedDestination);
+                }
             }
             return FileVisitResult.CONTINUE;
         }
